+++ lib/api/apiUtils/bucket/bucketEncryption.js
@@ -1,6 +1,5 @@
 const { errors } = require('arsenal');
 const { parseString } = require('xml2js');
-const vault = require('../../../auth/vault');
 
 /**
  * parseEncryptionXml - Parses and validates a ServerSideEncryptionConfiguration xml document
@@ -109,31 +108,7 @@ function parseBucketEncryptionHeaders(headers) {
     };
 }
 
-function checkExpectedBucketOwner(headers, bucket, log, cb) {
-    const expectedOwner = headers['x-amz-expected-bucket-owner'];
-    if (expectedOwner === undefined) {
-        return cb();
-    }
-
-    const bucketOwner = bucket.getOwner();
-    return vault.getAccountIds([bucketOwner], log, (error, res) => {
-        if (error) {
-            log.error('error fetch accountId from vault', {
-                method: 'checkExpectedBucketOwner',
-                error,
-            });
-        }
-
-        if (error || res[bucketOwner] !== expectedOwner) {
-            return cb(errors.AccessDenied);
-        }
-
-        return cb();
-    });
-}
-
 module.exports = {
     parseEncryptionXml,
     parseBucketEncryptionHeaders,
-    checkExpectedBucketOwner,
 };
