+++ lib/api/bucketDeleteEncryption.js
@@ -4,7 +4,7 @@ const metadata = require('../metadata/wrapper');
 const { metadataValidateBucket } = require('../metadata/metadataUtils');
 const { pushMetric } = require('../utapi/utilities');
 const collectCorsHeaders = require('../utilities/collectCorsHeaders');
-const { checkExpectedBucketOwner } = require('./apiUtils/bucket/bucketEncryption');
+const { checkExpectedBucketOwner } = require('./apiUtils/authorization/bucketOwner');
 
 /**
  * Bucket Delete Encryption - Delete bucket SSE configuration
