+++ lib/api/bucketGetEncryption.js
@@ -1,14 +1,11 @@
 const { errors, s3middleware } = require('arsenal');
+const async = require('async');
 
-const bucketShield = require('./apiUtils/bucket/bucketShield');
-const { isBucketAuthorized } =
-    require('./apiUtils/authorization/permissionChecks');
-const metadata = require('../metadata/wrapper');
 const { pushMetric } = require('../utapi/utilities');
-const escapeForXml = s3middleware.escapeForXml;
 const collectCorsHeaders = require('../utilities/collectCorsHeaders');
-
-const requestType = 'bucketGetEncryption';
+const { checkExpectedBucketOwner } = require('./apiUtils/authorization/bucketOwner');
+const { metadataValidateBucket } = require('../metadata/metadataUtils');
+const escapeForXml = s3middleware.escapeForXml;
 
 /**
  * Bucket Get Encryption - Get bucket SSE configuration
@@ -21,54 +18,41 @@ const requestType = 'bucketGetEncryption';
 
 function bucketGetEncryption(authInfo, request, log, callback) {
     const bucketName = request.bucketName;
-    const canonicalID = authInfo.getCanonicalID();
 
-    return metadata.getBucket(bucketName, log, (err, bucket) => {
-        if (err) {
-            log.debug('metadata getbucket failed', { error: err });
-            return callback(err);
-        }
-        if (bucketShield(bucket, requestType)) {
-            return callback(errors.NoSuchBucket);
-        }
-        log.trace('found bucket in metadata', {
-            bucket: bucketName,
-            method: 'bucketGetEncryption',
-        });
+    const metadataValParams = {
+        authInfo,
+        bucketName,
+        requestType: 'bucketGetEncryption',
+    };
 
-        const corsHeaders = collectCorsHeaders(request.headers.origin,
-            request.method, bucket);
-
-        if (!isBucketAuthorized(bucket, requestType, canonicalID, authInfo,
-        log)) {
-            log.debug('access denied for account on bucket', {
-                requestType,
-                method: 'bucketGetEncryption',
-            });
-            return callback(errors.AccessDenied, null, corsHeaders);
-        }
-
-        const sseInfo = bucket.getServerSideEncryption();
-
-        // If sseInfo is present but the `mandatory` flag is not set
-        // then this info was not created using bucketPutEncryption
-        // or by using the x-amz-scal-server-side-encryption header at
-        // bucket creation and should not be returned
-        if (sseInfo === null || !sseInfo.mandatory) {
-            log.trace('no server side encryption config found', {
-                bucket: bucketName,
-                method: 'bucketGetEncryption',
-            });
-            return callback(errors.ServerSideEncryptionConfigurationNotFoundError);
+    return async.waterfall([
+        next => metadataValidateBucket(metadataValParams, log, next),
+        (bucket, next) => checkExpectedBucketOwner(request.headers, bucket, log, err => next(err, bucket)),
+        (bucket, next) => {
+            const sseInfo = bucket.getServerSideEncryption();
+            if (sseInfo === null || !sseInfo.mandatory) {
+                log.trace('no server side encryption config found', {
+                    bucket: bucketName,
+                    method: 'bucketGetEncryption',
+                });
+                return next(errors.ServerSideEncryptionConfigurationNotFoundError);
+            }
+            return next(null, bucket, sseInfo);
+        },
+    ],
+    (error, bucket, sseInfo) => {
+        const corsHeaders = collectCorsHeaders(request.headers.origin, request.method, bucket);
+        if (error) {
+            return callback(error, corsHeaders);
         }
 
-        const xml = [];
-        xml.push(
+        const xml = [
             '<?xml version="1.0" encoding="UTF-8"?>',
             '<ServerSideEncryptionConfiguration>',
             '<Rule>',
             '<ApplyServerSideEncryptionByDefault>',
-            `<SSEAlgorithm>${escapeForXml(sseInfo.algorithm)}</SSEAlgorithm>`);
+            `<SSEAlgorithm>${escapeForXml(sseInfo.algorithm)}</SSEAlgorithm>`,
+        ];
 
         if (sseInfo.configuredMasterKeyId) {
             xml.push(`<KMSMasterKeyID>${escapeForXml(sseInfo.configuredMasterKeyId)}</KMSMasterKeyID>`);
