+++ lib/api/bucketPutEncryption.js
@@ -1,6 +1,7 @@
 const async = require('async');
 
 const { parseEncryptionXml } = require('./apiUtils/bucket/bucketEncryption');
+const { checkExpectedBucketOwner } = require('./apiUtils/authorization/bucketOwner');
 const metadata = require('../metadata/wrapper');
 const { metadataValidateBucket } = require('../metadata/metadataUtils');
 const kms = require('../kms/wrapper');
@@ -27,6 +28,7 @@ function bucketPutEncryption(authInfo, request, log, callback) {
 
     return async.waterfall([
         next => metadataValidateBucket(metadataValParams, log, next),
+        (bucket, next) => checkExpectedBucketOwner(request.headers, bucket, log, err => next(err, bucket)),
         (bucket, next) => {
             log.trace('parsing encryption config', { method: 'bucketPutEncryption' });
             return parseEncryptionXml(request.post, log, (err, encryptionConfig) => {
